Revert back two working version

+++ CIS411_Project/CIS411_Project.Core/Models/Category.cs
@@ -7,7 +7,7 @@ namespace CIS411_Project.Core.Models
 {
     public class Category
     {
-        public decimal CATEGORY_ID { get; set; }
+        public int CATEGORY_ID { get; set; }
 
         public string CATEGORY_NAME { get; set; }
     }
